Windows Acceptance Tests (#284)
This PR gets acceptance tests and screenshots working on windows.
To do that, a number of bugs caused by windows style filepaths were fixed. These include
* parsing the component name in the UI
* HMR not working on windows
* saving to file not working to windows if any new components were added via the UI

The windows and darwin (aka macos) screenshots have been separated into different folders.

For some context, windows supports both back and forward slashes as path separators, while unix only supports
forward slashes. However, nodejs on windows has a preference for using forward slashes, which can lead to issues for things like validation that checks for `filepath.startsWith(otherFilepath)` in the front end.
This is because the frontend uses `path-browserify` to do filepath related parsing, which is a version of the nodejs `path` module that is capable of running in the browser. (You cannot import the path module in the browser because it only exists in the nodejs environment)
path-browserify only supports running in posix mode, and not win32 mode. Which means that it does not work properly with windows path separators, and leads to the aforementioned issues with filepath.startsWith, because we often have scenarios where some filepaths are sent from studio-plugin which will use windows style paths and also generate some filepaths in the browser using path-browserify.

Originally, I wanted to figure out a way so that studio-plugin can use the native preferred path separators,
we can have a transformation layer to the frontend to change them all to posix path separators, and then we can turn
them back to native separators when sending information. However this turned out to be more complex than I expected,
so I think the simplest solution is to try to always deal with posix style paths. The difficulty here is that this is more error prone
than a bespoke layer, and can only be remedied by proper test coverage. On the other hand, we're pretty good about getting
proper test coverage on our code, so after going back and forth this is where I landed.
To do this, at first I tried importing path like `import path from "path/posix"`, which the nodejs docs said would ensure that we always use the posix version of the path api. However this used TOO much of the posix version of the api, and would change absolute paths like `C:/Users/oli`  to `/Users/oli`, which isn't a valid path on windows.
Instead, I replaced all usages of path in studio-plugin with [upath](https://www.npmjs.com/package/upath), which is a drop-in replacement for regular nodejs path that auto normalizes all paths.

This PR also fixes some windows specific issues around acceptance tests,
and splits the add-remove-static-page test into two separate tests.

I had to add shell: true to our spawn calls to get them to work in windows.
This isn't super well documented but a number of threads like the below exist about this.
https://stackoverflow.com/questions/60386867/node-spawn-child-process-not-working-in-windows

J=SLAP-2781
TEST=manual,auto

tested the above bugs are fixed

see that new screenshots will have their windows versions uploaded from CI