Fix Tooltips for Prop Docs Sometimes Not Appearing (#287)
This pull request fixes the bug where identical prop names in two different arrays or objects would result in only one of the tooltips appearing because the tooltip is anchored to the prop name. To solve this problem, this PR also adds the name of the object/array to the `anchorId` for the tooltip. 

J=SLAP-2822
TEST=auto