Hide Repeaters in UI (#290)
Hide the concept of Repeaters from the UI, except for the cases where the developer has created a Repeater in the code. These Repeaters will continue to be parsed and previewed as before, but the Repeater settings will not be editable through the UI. Instead of a toggle and text input appearing, a message bubble will be displayed (see screenshot).

J=SLAP-2828
TEST=auto, manual

In the test-site, see that the Repeater panel is gone for all components except for the repeated Banner. The prop panel for the repeated Banner looks like this:

<img width="325" alt="Screen Shot 2023-07-25 at 11 38 30 AM" src="https://github.com/yext/studio-prototype/assets/88398086/6b011e91-dd30-4a58-81c2-f4a179db4f53">