Use Vite's open config option (#299)
This PR removes the manual call to open the browser when starting the Studio server. Instead, we can use Vite's `open` config option. This allows us to remove `strictPort` and the server will open on the next available port if the specified one is already in use.

J=SLAP-2763
TEST=manual

See that serving the test-site opens the browser as expected and when 5173 is occupied, 5174 is used. See that Playwright tests run locally without opening the browser.