Fix z-index UI bugs (#298)
This PR fixes a couple bugs related to the z-index. Tooltips now appear above the active component highlighting and preview panel content, as does the red border that appears on hover for an error component preview.

This also partially fixes a recent regression in the field picker dropdown where the dropdown would get cut off and hidden in the panel for the ActiveComponentPropEditors because the y-overflow is hidden. The dropdown is now fixed on the page so it's visible and clickable. But another [item](https://yexttest.atlassian.net/browse/SLAP-2859) was made to fully fix the UI, because the dropdown now stays fixed on the page while scrolling.

J=SLAP-2846
TEST=auto, manual

Check tooltips and field picker dropdown in the test-site to make sure they are visible above the other page content.
Add a Playwright screenshot for the field picker dropdown to prevent future regressions.