Generate Test Data with Multiple New Pages (#304)
This PR updates the generate-test-data call that is made after creating a new page so that it will not wipe out
test data generated from pages that are not yet saved.
Previously, when we called studio's generate-test-data endpoint (which is a proxy to calling the generate-test-data CLI command) upon creating a new page, we could pass into the endpoint only the `StreamScope` for the new page, and combine that information with the features.json on file.
Now, instead, we pass an entire features.json to the endpoint. This features.json is created entirely in memory by combining the StreamScopes for all pages instead of just the newly created one.

When generate-test-data is run, it will generate new test data, and create a new mapping.json file that lists all of the test data. Previously, this file would be imported directly via a dynamic import. However, in this PR we now send the new mapping.json in the response of the generate-test-data endpoint. The issue with dynamic importing the mapping.json is that vite would cache the result after the first import, and as far as I can tell there is no way to invalidate this cache, even with things like `server.moduleGraph.invalidateModule`. I believe the reason for this is that the `ModuleNode.importers` property is always equal to an empty Set, which prevents vite from doing HMR for it out of the box, and we would have to write our own custom HMR if we wanted to do a dynamic import for it.

Finally, this PR fixes a bug where there's a chance that calling generate-test-data causes a full page refresh and wipes out any unsaved changes. This full page refresh was being caused by the CLI wiping out the contents of the localData folder before adding in newly generated test data. 
Full explanation at the bottom. 
tldr; we need to ignore the localData from being watched by adding it to the server.watch config.

J=SLAP-2849
TEST=manual,auto

I can create pages for entity type ids of ce_person and faq and see new localData, and also see that localData in the FieldPicker in the frontend

----

### HMR exploration

When this full page refresh happens, I noticed that vite would always output a logging message (in green) of "page reload", so I dug through the vite source code and looked for instances of this logging message.
This lead me to: [vite/src/node/server/hmr.ts](https://github.com/vitejs/vite/blob/f5c1224150fc8bfb2549e5c46d6c228776f8c170/packages/vite/src/node/server/hmr.ts) , which I found the corresponding chunk for in our node_modules and adding some logging.
What was happening is that, whenever a test data .json file was deleted, the updateModules() function below would be run, and `hasDeadEnd` would be set to true, which would trigger a full page reload.
![image](https://github.com/yext/studio-prototype/assets/23005393/14d5d9fe-b0c4-4612-83f2-4d20b228b77b)
`hasDeadEnd` would get set to true, because in propagateUpdate, if a ModuleNode has no importers, it will return true.
![image](https://github.com/yext/studio-prototype/assets/23005393/e1ea8cf0-6183-49d8-a921-dbf08d4da36b)
This kinda makes sense because if a module dependency gets updated but somehow has no importers something weird is going on and a page reload is a safe way to ensure correct behavior. [This is also almost 3 years old code](https://github.com/vitejs/vite/commit/0974dbeb7f4de03ff7c6091ca1a507470a0f6977)

Now the reason this can be solved by ignoring the localData from server.watch is that, the above code is called by handleFileAddUnlink within hmr.ts
![image](https://github.com/yext/studio-prototype/assets/23005393/90113624-ea30-4619-aac4-a93ccfc75010)

This is a listener handler, which is called here in [src/node/server/index.ts](https://github.com/vitejs/vite/blob/f5c1224150fc8bfb2549e5c46d6c228776f8c170/packages/vite/src/node/server/index.ts#L545), which is passed to a chokidar watcher, and that watcher is what takes in the config for server.watch.

I believe we may run into future caching issues for test data similar to the mapping.json, where preexisting test data is updated by the CLI and the data has changed, but the frontend doesn't see the changes, but that's not something I've tested in this PR.