Always sync entity files after generating test data (#326)
This PR updates our test data generation flow so that even if the generation returns a non-zero exit code, we still pass the resulting `mapping.json` to studio. The store is updated based on the new state of the `localData` directory. Also, toast messages were changed from a warning to an error because currently the main use case for Studio is for developing dynamic entity templates, so a problem in the test data is a major issue that should likely result in the user exiting out of the UI.

This PR also updates the test-site's `localData` script so it pulls all test data fields with the `-a` flag, and it fixes a couple bugs
- When `entityFiles` for the active page are undefined, `refreshActivePageEntities` changes the `activePageEntities` to undefined, instead of just returning without making any updates
- When the stream scope for a non-active page is updated and test data is regenerated, we don't don't update the `activeEntityFile` unless the test data for the active page has been changed. In that case, one of the active page's `entityFiles` is set as the new `activeEntityFile`, not an entity file from the edited page that triggered the test data generation

J=SLAP-2867
TEST=manual, auto

In the test-site, see that
- errored test data generation results in an error toast
- a new page with no test data after regeneration no longer shows the entity picker with entities from the previous active page
- regenerating test data when the stream scope for a non-active page is edited no longer gives a console error about incorrectly trying to update the activeEntityFile