Add Location Entity Template to E2E Test Site.  (#331)
This PR is the first step in making our `e2e-tests` a bit more realistic. We will have Templates and Components that are a bit more complex. Our end goal is for the `e2e-tests` repo to look and feel like a (somewhat) real PagesJS repo. This ensures that we are running our acceptance and visual regression tests in scenarios that are as real life as possible. Eventually, we will want to have a full end-to-end test that makes a change in Studio and sees it propagate through the
PagesJS `dev` and `build` commands.

For now, we add a more realistic Template for a Location Page, as well as some common Components for such a page. The `add-element` test was updated to add a Footer to the Location Template.

TEST=auto

Ran all unit tests and re-generated all the necessary snapshots. After re-generating the snapshots, I did spot check a good number of them. I checked all the snapshots for the modified `add-element` test.