Rework HMR to not use handleHotUpdate (#333)
This PR separates our custom HMR from vite's handleHotUpdate hook.
Instead, we now create our own watcher just for a studio user's src
files.
Overall, this should be more maintanable in the future, since vite's
default HMR does not work well for a studio user's src files due to the
atypical way we're using using them.

We also now fire HMR updates when a user updates the body of a
component, which is behavior we had at one point but lost somewhere
along the way. A test was added for this behavior.

The issue I was running into was that, after adding a custom tailwind
config in a [separate PR](https://github.com/yext/studio/pull/323),
whenever a page was deleted from the filesystem, studio would do a full
page refresh. This was occurring due to tailwind registering the user's
src files as dependencies
(specifically right here
https://github.com/tailwindlabs/tailwindcss/blob/ff2637257070c2ce8584603424d0063eba422b27/src/lib/setupTrackingContext.js#L126).
This would cause the file to be hooked up to vite's 'unlnk' listener 

https://github.com/vitejs/vite/blob/77175099489ffbabefb9d98d448c782ced2986b5/packages/vite/src/node/server/index.ts#L560
The behavior of the unlink listener skips the handleHotUpdate hook and
cannot be customized. When this same issue popped up in
https://github.com/yext/studio/pull/304 for the localData folder, we
could simply add the localData folder to the watcher's ignore options.
We can't do that for a user's page templates, because completely
ignoring them would mean we can't perform HMR updates on them.

The new architecture imo is simpler to understand because it only deals
with studio specific details instead of dealing with all vite hot
updates (which we don't need to touch since they work just the way we'd
like and we're not doing anything weird outside of how we use the user's
src files).

J=SLAP-2883,SLAP-2607
TEST=manual,auto

tested hot module reloading works for editing a page and editing a
component
I can delete a page via UI and hit save and there is no full page reload
tests for adding and removing new components and pages

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>