Components: Allow field picker to scroll with page (#330)
This fixes an issue where the field picker dropdown would not scroll
when the page scrolled. This also fixes an issue with the prop editor
menu button appearing outside of the visible area.

J=SLAP-2859
TEST=auto,manual

Checked that the field picker dropdown scrolls appropriately and that
the prop editor menu button appears next to the ellipses icon for all
sidebar widths as expected. Also added a test for the menu button.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>