Add URL Slug Validation (#291)
For both entity and static pages, this pull request creates a function
to check whether getPath() returns a valid or invalid (undefined or
containing unsafe characters) URL slug value. If it's invalid, the page
settings modal does not allow editing of the URL slug. If it's valid,
then the modal does allow editing. However, if the user tries to save an
invalid URL slug in the modal, then an error message is displayed and
the Save button is disabled until new user input.

For an entity page, a valid URL slug also includes any string that
contains ${document.field} where "field" is variable.

J=SLAP-2826
TEST=auto, manual

Checked that URL behavior works as expected in the test site.

---------

Co-authored-by: github-actions <41898282+github-actions[bot]@users.noreply.github.com>
Co-authored-by: Oliver Shi <oshi@yext.com>
Co-authored-by: Jacob <115099870+jwartofsky-yext@users.noreply.github.com>
Co-authored-by: Jacob Wartofsky <jwartofsky@yext.com>