Set up requests to the Management API (#350)
This PR sets up requests to the Management API. Because the Management API doesn't currently pass the `Access-Control-Allow-Origin` header with the response, we're not able to make requests directly from the browser. Instead, `studio` will need to send a message to `studio-plugin` to fire off those requests using `cross-fetch`.

Also, because we don't know how we'll get the user's API key yet, this functionality is turned off for now (see [this](https://github.com/yext/studio/pull/350/commits/a4bba34709b8ec6de1d6c26773f219214906f5d7) commit for a partial example of how it would be added). Lastly, this PR fixes a bug in `registerListener` so that it correctly awaits the `handler` before returning. This way, a rejected promise is actually caught by the `catch` block.

Because there's discussion around updating the Management API to pass the correct CORS header in the response, we're hopeful that this code can soon be updated so the requests are made directly from the browser. To make that switch easier, I've set up the message handling so there is a message corresponding with each API endpoint method. This way, most of the data aggregation logic will still live in `studio`, rather than `studio-plugin`. This means we're sending some extra messages to `studio-plugin`, but I did some experimenting and the majority of the time to aggregate the data is due to the API requests rather than the message passing within Studio, so there won't be too much of a performance difference (the extra time is ~350ms for the Slapshot Test account). The plan is to have most of this aggregation occur asynchronously on start up of the Studio store, so it shouldn't affect the load time of the stream scope modal.

J=SLAP-2861
TEST=manual

Manually supply an API key and see that I can make requests from Studio to get all the information we need to populate the stream scope modal.