Add prominent image to location page (#354)
Adds a prominent image to the `e2e-tests` site.  The image we currently use in the test site is 16 MB, and would not load in time for the automated tests.  A smaller image (103 KB) of NYC was found and is used instead. Thank you, [Roberto Vivancos](https://www.pexels.com/@robertovivancos/), for the free-to-use image.

[Original image source.](https://www.pexels.com/photo/skyline-photo-of-empire-state-building-in-new-york-city-2190283/)
[Compressed image used in the test site.
](https://images.pexels.com/photos/2190283/pexels-photo-2190283.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=1)
[Can use commercially](https://help.pexels.com/hc/en-us/articles/360042295214-Can-I-use-the-photos-and-videos-for-a-commercial-project-).

J=SLAP-2889
TEST=manual