Use studio-plugin version in package.json instead of wildcard (#355)
In npm workspaces, you can use `*` to point to a local dependency (i.e.
another package), which is how we add `studio-plugin` as a dependency of
`studio`. This dependency is preserved as a symlink in the root
`package-lock.json`. Since the dependency version of `studio-plugin` in
the `studio` `package.json` was `*`, there was not a concrete version of
`studio-plugin` specified, so users had to explicitly install `studio`
and `studio-plugin`. This PR finds that we can still specify a version
of `studio-plugin` in the `studio` `package.json` and still preserve the
symlink for local development.

J-SLAP-2895
TEST=manual