Add dropdowns to stream scope modal (#359)
This PR updates the modal for specifying the stream scope in the add page flow so the inputs have a similar UI to the custom Tailwind class input. Each input will have an icon that opens a dropdown with the possible options that can be added for setting the scope. So far, the dropdowns for entity types and saved filters will be correctly populated from the store (tested with dummy data for now). The remaining subtasks will populate the store with actual account content and populate the dropdown for entities, which is more complex.

Also, make the disabled behavior for the Tailwind class prop input more robust: 1) the embed icon is no longer clickable when the input is disabled and 2) the disabled state is more visibly different from the enabled state.

J=SLAP-2905, 2910
TEST=auto, manual

Check that a new page can be created with a stream scope with zero, one, or multiple filters. Makes sure that the disabled state works correctly so that selecting one type of filter (e.g. entity type) will disable the other inputs (saved filters and entities).