Populate initial account content (#366)
This PR populates the store with the account content for saved filters, entity types, and the first 50 entities of each type. The `PillPicker` dropdown styling is also updated to be a fixed max height with scrollable content, instead of growing the dropdown to display all options.

J=SLAP-2906
TEST=auto, manual

Hard-code an API key in studio-plugin and see the stream scope modal for adding an entity page in the test site is populated with the entity types and saved filters from the Slapshot test account. Entity types with no entities are not included in the entity types dropdown.