Introduce `Layouts` to Studio (#367)
This PR introduces Layouts to Studio. Studio now recognizes layouts that
appear in `src/layouts` (or a custom dir) and stores their filepaths in
a new `layouts` slice. Layout names are derived from the layout's
filename. Tests were updated to reflect changes in studio config. A test
was written to check that layout files are found from the `layouts` dir.

The current `LayoutSlice` just contains a `string[]` of the names of the
layouts in the layouts dir. This is simply a place holder, since no
layout parsing was implemented yet. Layout parsing and the LayoutSlice
will be fleshed out in future PRs in the epic.

J=SLAP-2923
TEST=auto,manual