Pass the injected `YEXT_STUDIO_API_KEY` to the `ManagementApiService`.
This PR adds logic to the creation of the Studio Vite Plugin. If there is
an env var called `YEXT_STUDIO_API_KEY`, we use that to create an instance
of the `ManagementApiService`. This environment variable should be available
in all Studio CBD containers. The value of the variable is an API Key that will
give the Studio instance access to the Management API. This is the API that
powers Stream Scope Selection.

TEST=manual

Tested locally. I did not have the env var set and saw that Stream Scope
Selection was disabled. Then I set the env var and saw that the Scope
Selection dropdowns were populated correctly.