Add support for parsing layouts (#370)
This PR creates a new `LayoutMetadata` domain model to represent a
Layout. A new `LayoutFile` sourcefile is added for parsing layouts.
Currently, we only support layouts in the top-level of the `src/layouts`
directory (i.e. no nested folders). Any nested directories for both
layout and pages will now be ignored.

I opted not to create a model for layouts with errors because, for now,
this information is not anticipated to be used in the UI. Like
components, we will not be showing layouts with errors in the Add
Element menu. Unlike components, there is nowhere else in the UI an
errored layout would surface, since it's not a "component" that would be
part of a component tree. The parsing error is still logged in the
terminal though.

J=SLAP-2929
TEST=auto

Since much of the logic is shared between parsing a page and parsing a
layout, I moved the tests for parsing a component tree into a separate
file so only sourcefile-specific tests remain in
`PageFile.getPageState.test.tsx`.

---------

Co-authored-by: Oliver Shi <oshi@yext.com>