Edit Stream Scope Improvements (#374)
This PR adds the add page flow stream scope improvements to the page
edits. There is now a new edit scope button on the page panel next to
the settings button which opens up a new Stream Scope Edit modal.

The modal is pre-populated by the current stream scope config in the
page template. Since we are only allowing users to edit one stream scope
at a time, if the developer has specified a mix of stream scopes, all
editing will be disabled. Users can only submit their new selection if
changes have been made, similar to the old page settings.

If there are no entity types or saved filters, then these fields will
not be editable. In these cases, even if the developer has specified a
stream scope, it will not be displayed.

J=SLAP-2918
TEST=auto, manual