Version 0.23.0. (#382)
## Features
- We've introduced the concept of Page Layouts (or Layouts) to Studio. A
Page Layout is a pre-prescribed set of Components that you can apply to
a Page. By default, Layouts live in the `src/layouts` directory. In the
UI, a Layout can be applied via the "Insert" button. (#370, #367, #375)
- There is now improved UX for creating or editing an Entity Page's
Stream Scope. Instead of being raw inputs, the fields for Entity Ids,
Entity Types, and Saved Filters are now drop-downs with the relevant
options populated from the account. (#380)

## Changes
- Repeaters and Modules have been officially deprecated and removed from
Studio. (#369)