Live Preview Button Opens the Corresponding Static/Entity Page. (#379)
This PR changes the behavior of the "Live Preview" button. Previously,
this button would open the PagesJS Landing Page in a new tab. Now, it
opens a PagesJS-powered preview of the active Page in Studio. If the
active Page is an Entity Page, the preview uses the currently selected
Entity in Studio. There are a couple scenarios in which the Live Preview
button will be disabled:

1. When we cannot properly determine the URL of active Page. 
2. If there is no active Page currently in Studio. 

In a later PR, I will add a Tooltip that appears when the button is
disabled. It will explain why it's disabled. In this PR, I also removed
the "Open Live Preview" Acceptance test. I don't think that test adds
much value any more. We have extensive unit test coverage for the Live
Preview button. This coverage ensures that, when enabled, `window.open`
is called with the correct URL. We can assume that `window.open` works
correctly. We already have various screenshots that capture the button
(in both the enabled and disabled state). So, we've already got visual
regression coverage.


TEST=auto,manual

Ensured all the new unit tests passed. Using the Test Site, I verified
that the proper Preview tab opened for both a Static and Entity page.