Create `ViewportMenuSection` to Take Scoped Screenshots of Viewport Menu (#390)
In this PR, I've created a `ViewportMenuSection` class. This class acts
as a Playwright wrapper around `ViewportButton` and `ViewportMenu`. We
can now offload the Viewport-specific code from `StudioPlaywrightPage`
to this new class. The new class has a `takeScreenshot` method which is
scoped only to the Viewport drop-down. I'm hoping that this PR
establishes a new pattern in designing our Playwright tests. We break up
the Studio app into different Playwright "Sections". Each Section
encapsulates the functionality for that piece of Studio and generates
appropriately-scoped screenshots.

J=SLAP-2922
TEST=auto

Regenerated snapshots for the `viewport-menu.spec.ts` tests. Saw they
were correctly scoped.