Create `AddElementSection` and Reduce Number of Screenshots. (#394)
This PR adds the `AddElementSection`, a Playwright wrapper for the "Add
Element" functionality in Studio. Logic for adding Elements was moved
from `StudioPlaywrightPage` into this new Section. The Section is also
able to take targeted screenshots.

Also in the PR, I reduced the number of screenshots we were taking. In
the "can update initial props of a component and see UI is updated via
HMR" test, we were taking 8 screenshots.

J=SLAP-2922
TEST=auto