Support inheritance when parsing interfaces (#395)
This PR adds support for inherited properties when parsing interfaces such as component prop interfaces or site settings. Something like this is now supported:
```ts
import { CtaData } from "@yext/search-ui-react";

interface BaseProps {
  title: string;
}

interface ComponentProps extends BaseProps, CtaData {
  description: string;
}
```
If we're not able to resolve the type that's being extended to an object type, we throw an error. This would include extending a non-object type like `HexColor` (which is invalid TS), extending a built-in type like `Record`, etc. We also aren't adding any additional support for type unions or intersections, beyond our current support for string unions, since TS does more complicated merging when properties are shared in those cases. For interfaces that inherit properties, if a property is shared between the inherited type and the new type, the new property's type is required by TS to be a subset of the inherited property's type, and the new property's type is used in the final, generated type. In other words, only type narrowing is allowed.

J=SLAP-2938
TEST=auto, manual

In the test-site, see that prop editors for inherited props show up as expected in the UI.