Fix imports for commonly-named components (#398)
This PR fixes imports for commonly-named components like `Container` and `Text` by manually writing their imports to the file before calling `ts-morph`s `fixMissingImports` function. Previously, this function would often add an incorrect import in the case of `Container` and not add any import in the case of `Text`, which is also a built-in interface in React.

J=SLAP-2587
TEST=auto, manual

Added `postcss` to the test site and saw that without these changes, `Container` was imported from `postcss` (which also has a named export called `Container`), and that with these changes, `Container` is correctly imported from `src/components`. Also saw that the behavior for `Text` was fixed (i.e. an import was actually added for a `Text` component rather than it trying to use the React type).