Remove redundant snapshots from 'Add Page Using a Layout' test. (#400)
This PR updates `add-page-with-layout.spec.ts` to use the new `AddPageSection`. In addition, I've removed redundant snapshots. We have snapshots of the modal flow for adding a static page w/out a Layout. So, here, we just need a snapshot of the Layout Modal. We also don't need a snapshot of the newly created page pre-save. We test that the "Save" button is enabled by clicking it. And we have visual regression tests already for the enabled "Save" button.

J=SLAP-2922
TEST=auto