Use `AddPageSection` in Playwright When Adding an Entity Page.  (#401)
This PR updates the `can-add-entity-page` test to use the `AddPageSection`. With that, all logic for adding a Static/Entity Page can finally be removed from `StudioPlaywrightPage`. I cut down the number of snapshots taken by this test significantly. It only takes snapshots pertaining to Scope Selection (the one part of the Add Page Flow we still need regression coverage for) and once the new Entity Page is saved.

I added a TODO. To further improve our visual regression test coverage, we should have the Scope Selection Modal actually be populated with data that we can expand. But, that's outside the scope of this PR.

J=SLAP-2922
TEST=auto