Add an Info Icon for Props with a Tooltip. (#403)
This PR adds an Info icon to the right of the Prop Name when the Prop has a Tooltip. The anchor of the Tooltip is now the Info Icon. To do this, I created a new common Component called TooltipIcon representing the combination of an `Icon` and a `Tooltip`. I added visual regression coverage for Prop Tooltips as well, we were missing it.

J=SLAP-2952
TEST=auto