Support the Omit utility type for a prop (#413)
This PR adds support for the `Omit` TypeScript utility type when defining the type of a prop in a component's prop interface. For now, utility types are still not allowed when defining the overall type of the component's prop interface (i.e. `props: Omit<MyProps, "blah">`).

This PR also adds a few new errors:
- An error for when we're unable to locate/resolve a component's prop interface (for example, if they use a utility type for the overall type as mentioned above). Previously, we would silently continue and assume there were no props for the component, which could throw obscure TS errors later on when parsing.
- An error for when a value was present for a field within an object prop on an instance of a component in a PageFile, but that field wasn't defined in the prop interface of that component. This would also throw an unclear TS error previously.
- An error for when an initial prop was present for a component that wasn't defined in its prop interface. This would also throw an unclear TS error previously.

J=SLAP-2967
TEST=auto, manual

In the test site, see that using `Omit<ObjectProp, "nestedBool">` instead of `ObjectProp` for the type of `obj` in `BannerData` correctly results in the prop editor for `nestedBool` not appearing in the UI.