Support the Pick utility type for a prop (#414)
This PR adds support for the `Pick` TypeScript utility type when defining the type of a prop in a component's prop interface. As mentioned in the previous PR, utility types are still not allowed when defining the overall type of the component's prop interface (i.e. `props: Pick<MyProps, "blah">`).

J=SLAP-2968
TEST=auto, manual

In the test site, see that using `Pick<ObjectProp, "nestedBool">` instead of `ObjectProp` for the type of `obj` in `BannerData` correctly results in the prop editor for `nestedBool` being the only one that appears in the UI for that prop.